% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIEDA.R
\name{runIEDA}
\alias{runIEDA}
\title{Run IEDA Shiny Application to perform Interactive Exploratory Data Analysis in the package IEDA}
\usage{
runIEDA()
}
\description{
Exploratory Data Analysis is one of the key components of a data science project.
It is a crucial step to take before diving in to the machine learning or data modeling.
EDA can sometimes be time consuming due to reasons like huge number of variables in the data set, writing code for each plot or group of plots.
In order to eliminate these difficulties and reduce the time here is a simple tool which provides different interactive visualizations used in EDA just on click of a button.
}
\examples{
runIEDA()
}
\keyword{Analysis}
\keyword{Data}
\keyword{EDA,}
\keyword{Exploratory}
\keyword{IEDA,}
